Add BoxWithClose, add approval step
