+++ components/delegations/DelegateModal.tsx
@@ -14,6 +14,7 @@ import TxIndicators from '../TxIndicators';
 import useAccountsStore from '../../stores/accounts';
 import useTransactionStore, { transactionsSelectors, transactionsApi } from '../../stores/transactions';
 import MKRInput from '../MKRInput';
+import { BoxWithClose } from 'components/BoxWithClose';
 
 type Props = {
   isOpen: boolean;
@@ -47,7 +48,18 @@ export default function DelegateModal({ isOpen, onDismiss, delegate }: Props): J
   // TODO: add approval tx and state
   const approveMkr = async () => {
     const maker = await getMaker();
-    const tx = maker.getToken(MKR).approveUnlimited(delegate.address);
+    const approveTxCreator = () => maker.getToken(MKR).approveUnlimited(delegate.address);
+    const txId = await track(approveTxCreator, 'Approving MKR', {
+      mined: txId => {
+        transactionsApi.getState().setMessage(txId, 'MKR approved');
+        setTxId(null);
+      },
+      error: () => {
+        transactionsApi.getState().setMessage(txId, 'MKR approval failed');
+        setTxId(null);
+      }
+    });
+    setTxId(txId);
   };
 
   const lockMkr = async () => {
@@ -57,11 +69,11 @@ export default function DelegateModal({ isOpen, onDismiss, delegate }: Props): J
     const txId = await track(lockTxCreator, 'Depositing MKR', {
       mined: txId => {
         transactionsApi.getState().setMessage(txId, 'MKR deposited');
-        onDismiss();
+        // onDismiss();
       },
       error: () => {
         transactionsApi.getState().setMessage(txId, 'MKR deposit failed');
-        onDismiss();
+        // onDismiss();
       }
     });
     setTxId(txId);
@@ -90,101 +102,115 @@ export default function DelegateModal({ isOpen, onDismiss, delegate }: Props): J
                 }
           }
         >
-          {tx ? (
-            <Flex sx={{ flexDirection: 'column', textAlign: 'center' }}>
-              <Text variant="microHeading" color="onBackgroundAlt">
-                {txPending ? 'Transaction pending' : 'Confirm transaction'}
-              </Text>
+          <BoxWithClose
+            content={
+              <Box>
+                {tx ? (
+                  <Flex sx={{ flexDirection: 'column', textAlign: 'center' }}>
+                    <Text variant="microHeading" color="onBackgroundAlt">
+                      {txPending ? 'Transaction pending' : 'Confirm transaction'}
+                    </Text>
 
-              <Flex sx={{ justifyContent: 'center', mt: 4 }}>
-                <TxIndicators.Pending sx={{ width: 6 }} />
-              </Flex>
-
-              {!txPending && (
-                <Box sx={{ mt: 4 }}>
-                  <Text sx={{ color: 'mutedAlt', fontSize: 3 }}>
-                    Please use your wallet to confirm this transaction.
-                  </Text>
-                  <Text
-                    sx={{ color: 'muted', cursor: 'pointer', fontSize: 2, mt: 2 }}
-                    onClick={() => setTxId(null)}
-                  >
-                    Cancel
-                  </Text>
-                </Box>
-              )}
-            </Flex>
-          ) : (
-            <>
-              {mkrBalance && hasLargeMkrAllowance ? (
-                <Flex sx={{ flexDirection: 'column', alignItems: 'center', justifyContent: 'center' }}>
-                  <Text variant="microHeading" sx={{ fontSize: [3, 6] }}>
-                    Deposit into delegate contract
-                  </Text>
-                  <Text sx={{ color: 'secondaryEmphasis', mt: 2 }}>
-                    Input the amount of MKR to deposit into the delegate contract.
-                  </Text>
-                  <Box sx={{ mt: 3, width: '20rem' }}>
-                    <Flex sx={{ border: '1px solid #D8E0E3', justifyContent: 'space-between' }}>
-                      <MKRInput
-                        onChange={setMkrToDeposit}
-                        placeholder="0.00 MKR"
-                        error={mkrToDeposit.gt(mkrBalance) && 'MKR balance too low'}
-                        style={{ border: '0px solid', width: bpi < 1 ? '100%' : null, m: 0 }}
-                        ref={input}
-                      />
-                      <Button
-                        disabled={mkrBalance === undefined}
-                        variant="textual"
-                        sx={{ width: '100px', fontWeight: 'bold' }}
-                        onClick={() => {
-                          if (!input.current || mkrBalance === undefined) return;
-                          changeInputValue(input.current, mkrBalance.toBigNumber().toString());
-                        }}
-                      >
-                        Set max
-                      </Button>
+                    <Flex sx={{ justifyContent: 'center', mt: 4 }}>
+                      <TxIndicators.Pending sx={{ width: 6 }} />
                     </Flex>
-                    <Flex sx={{ alignItems: 'baseline', mb: 3, alignSelf: 'flex-start' }}>
-                      <Text
-                        sx={{
-                          textTransform: 'uppercase',
-                          color: 'secondaryEmphasis',
-                          fontSize: 1,
-                          fontWeight: 'bold'
-                        }}
-                      >
-                        MKR Balance:&nbsp;
-                      </Text>
-                      {mkrBalance ? (
-                        <Text
-                          sx={{ cursor: 'pointer', fontSize: 2, mt: 2 }}
-                          onClick={() => {
-                            if (!input.current) return;
-                            changeInputValue(input.current, mkrBalance.toBigNumber().toString());
-                          }}
-                        >
-                          {mkrBalance.toBigNumber().toFormat(6)}
+
+                    {!txPending && (
+                      <Box sx={{ mt: 4 }}>
+                        <Text sx={{ color: 'mutedAlt', fontSize: 3 }}>
+                          Please use your wallet to confirm this transaction.
                         </Text>
-                      ) : (
-                        <Box sx={{ width: 6 }}>
-                          <Skeleton />
+                        <Text
+                          sx={{ color: 'muted', cursor: 'pointer', fontSize: 2, mt: 2 }}
+                          onClick={() => setTxId(null)}
+                        >
+                          Cancel
+                        </Text>
+                      </Box>
+                    )}
+                  </Flex>
+                ) : (
+                  <>
+                    {mkrBalance && hasLargeMkrAllowance ? (
+                      <Flex sx={{ flexDirection: 'column', alignItems: 'center', justifyContent: 'center' }}>
+                        <Text variant="microHeading" sx={{ fontSize: [3, 6] }}>
+                          Deposit into delegate contract
+                        </Text>
+                        <Text sx={{ color: 'secondaryEmphasis', mt: 2 }}>
+                          Input the amount of MKR to deposit into the delegate contract.
+                        </Text>
+                        <Box sx={{ mt: 3, width: '20rem' }}>
+                          <Flex sx={{ border: '1px solid #D8E0E3', justifyContent: 'space-between' }}>
+                            <MKRInput
+                              onChange={setMkrToDeposit}
+                              placeholder="0.00 MKR"
+                              error={mkrToDeposit.gt(mkrBalance) && 'MKR balance too low'}
+                              style={{ border: '0px solid', width: bpi < 1 ? '100%' : null, m: 0 }}
+                              ref={input}
+                            />
+                            <Button
+                              disabled={mkrBalance === undefined}
+                              variant="textual"
+                              sx={{ width: '100px', fontWeight: 'bold' }}
+                              onClick={() => {
+                                if (!input.current || mkrBalance === undefined) return;
+                                changeInputValue(input.current, mkrBalance.toBigNumber().toString());
+                              }}
+                            >
+                              Set max
+                            </Button>
+                          </Flex>
+                          <Flex sx={{ alignItems: 'baseline', mb: 3, alignSelf: 'flex-start' }}>
+                            <Text
+                              sx={{
+                                textTransform: 'uppercase',
+                                color: 'secondaryEmphasis',
+                                fontSize: 1,
+                                fontWeight: 'bold'
+                              }}
+                            >
+                              MKR Balance:&nbsp;
+                            </Text>
+                            {mkrBalance ? (
+                              <Text
+                                sx={{ cursor: 'pointer', fontSize: 2, mt: 2 }}
+                                onClick={() => {
+                                  if (!input.current) return;
+                                  changeInputValue(input.current, mkrBalance.toBigNumber().toString());
+                                }}
+                              >
+                                {mkrBalance.toBigNumber().toFormat(6)}
+                              </Text>
+                            ) : (
+                              <Box sx={{ width: 6 }}>
+                                <Skeleton />
+                              </Box>
+                            )}
+                          </Flex>
+                          <Button onClick={lockMkr} disabled={txPending} sx={{ width: '100%' }}>
+                            Delegate MKR
+                          </Button>
                         </Box>
-                      )}
-                    </Flex>
-                    <Button onClick={lockMkr} disabled={txPending} sx={{ width: '100%' }}>
-                      Delegate MKR
-                    </Button>
-                    {/* <Flex sx={{ justifyContent: 'center' }}>
-                  <TxIndicators.Pending sx={{ width: 6 }} />
-                </Flex> */}
-                  </Box>
-                </Flex>
-              ) : (
-                <Box>TODO: Approval content</Box>
-              )}
-            </>
-          )}
+                      </Flex>
+                    ) : (
+                      <Flex sx={{ flexDirection: 'column', textAlign: 'center' }}>
+                        <Text variant="microHeading" color="onBackgroundAlt">
+                          Approve Delegate Contract
+                        </Text>
+                        <Text sx={{ color: 'secondaryEmphasis', mt: 3 }}>
+                          Approve the transfer of MKR tokens to the delegate contract to deposit your MKR.
+                        </Text>
+                        <Button onClick={approveMkr} disabled={txPending} sx={{ width: '100%', mt: 3 }}>
+                          Approve Delegate Contract
+                        </Button>
+                      </Flex>
+                    )}
+                  </>
+                )}
+              </Box>
+            }
+            close={onDismiss}
+          />
         </DialogContent>
       </DialogOverlay>
     </>
