+++ components/executive/Withdraw.tsx
@@ -1,6 +1,6 @@
 /** @jsx jsx */
 import { useState, useRef } from 'react';
-import { Button, Flex, Text, Close, Box, jsx, Card, Alert } from 'theme-ui';
+import { Button, Flex, Text, Box, jsx, Alert } from 'theme-ui';
 import { DialogOverlay, DialogContent } from '@reach/dialog';
 import { useBreakpointIndex } from '@theme-ui/match-media';
 import Skeleton from 'react-loading-skeleton';
@@ -18,6 +18,7 @@ import useTransactionStore, { transactionsSelectors, transactionsApi } from 'sto
 import { changeInputValue } from 'lib/utils';
 import invariant from 'tiny-invariant';
 import mixpanel from 'mixpanel-browser';
+import { BoxWithClose } from 'components/BoxWithClose';
 
 const ModalContent = ({ address, voteProxy, close, ...props }) => {
   invariant(address);
@@ -205,25 +206,6 @@ const ModalContent = ({ address, voteProxy, close, ...props }) => {
 
   return <BoxWithClose content={content} close={close} {...props} />;
 };
-
-export const BoxWithClose = ({ content, close, ...props }): JSX.Element => (
-  <Box sx={{ position: 'relative' }} {...props}>
-    <Close
-      aria-label="close"
-      sx={{
-        height: 4,
-        width: 4,
-        position: 'absolute',
-        top: '-17px',
-        right: '-42px',
-        display: ['none', 'block'],
-        outline: 'none'
-      }}
-      onClick={close}
-    />
-    {content}
-  </Box>
-);
 
 const Withdraw = (props): JSX.Element => {
   const account = useAccountsStore(state => state.currentAccount);
