+++ components/executive/WithdrawOldChief.tsx
@@ -1,11 +1,11 @@
 /** @jsx jsx */
-import { useState, useRef } from 'react';
-import { Button, Flex, Text, Close, Box, jsx, Card, Alert } from 'theme-ui';
+import { useState } from 'react';
+import { Button, Flex, Text, Box, jsx, Alert } from 'theme-ui';
 import { DialogOverlay, DialogContent } from '@reach/dialog';
 import { useBreakpointIndex } from '@theme-ui/match-media';
 import shallow from 'zustand/shallow';
 import useSWR from 'swr';
-import Stack from '../layouts/Stack';
+import Stack from 'components/layouts/Stack';
 import getMaker, { MKR, getNetwork } from 'lib/maker';
 import useAccountsStore from 'stores/accounts';
 import { CurrencyObject } from 'types/currency';
@@ -18,6 +18,7 @@ import oldChiefAbi from 'lib/abis/oldChiefAbi.json';
 import oldVoteProxyAbi from 'lib/abis/oldVoteProxyAbi.json';
 import oldIouAbi from 'lib/abis/oldIouAbi.json';
 import { oldChiefAddress, oldIouAddress } from 'lib/constants';
+import { BoxWithClose } from 'components/BoxWithClose';
 
 const ModalContent = ({ address, voteProxy, close, ...props }) => {
   invariant(address);
@@ -187,25 +188,6 @@ const ModalContent = ({ address, voteProxy, close, ...props }) => {
 
   return <BoxWithClose content={content} close={close} {...props} />;
 };
-
-export const BoxWithClose = ({ content, close, ...props }): JSX.Element => (
-  <Box sx={{ position: 'relative' }} {...props}>
-    <Close
-      aria-label="close"
-      sx={{
-        height: 4,
-        width: 4,
-        position: 'absolute',
-        top: '-17px',
-        right: '-42px',
-        display: ['none', 'block'],
-        outline: 'none'
-      }}
-      onClick={close}
-    />
-    {content}
-  </Box>
-);
 
 const WithdrawOldChief = (props): JSX.Element => {
   const account = useAccountsStore(state => state.currentAccount);
